+++ CHANGELOG.md
@@ -86,6 +86,7 @@
     - read-only symbol/field access
     - no-return expression checking in finally-block
     - illegal concatenation of parameter expansion
+    - only available in global scope
 - preserve exit status during the evaluation of finally-block
 - complete type template, such as ``Array``, ``Map``, ``Tuple``, ``Option``
 - do not complete methods that do not satisfy type constraints